+++ tests/LodgingServiceFactoryTest.php
@@ -4,6 +4,7 @@ namespace DigipolisGent\Tests\Toerismevlaanderen\Lodging;
 
 use DigipolisGent\API\Client\ClientInterface;
 use DigipolisGent\Toerismevlaanderen\Lodging\Handler\CountHandler;
+use DigipolisGent\Toerismevlaanderen\Lodging\Handler\ListHandler;
 use DigipolisGent\Toerismevlaanderen\Lodging\LodgingService;
 use DigipolisGent\Toerismevlaanderen\Lodging\LodgingServiceFactory;
 use PHPUnit\Framework\TestCase;
@@ -22,6 +23,7 @@ class LodgingServiceFactoryTest extends TestCase
     {
         $clientMock = $this->prophesize(ClientInterface::class);
         $clientMock->addHandler(new CountHandler())->shouldBeCalled();
+        $clientMock->addHandler(new ListHandler())->shouldBeCalled();
 
         $service = LodgingServiceFactory::create($clientMock->reveal());
         $this->assertInstanceOf(LodgingService::class, $service);