+++ tests/Request/ListRequestTest.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace DigipolisGent\Tests\Toerismevlaanderen\Lodging\Request;
+
+use DigipolisGent\Toerismevlaanderen\Lodging\Filter\LocalityFilter;
+use DigipolisGent\Toerismevlaanderen\Lodging\Filter\RegistrationStatusFilter;
+use DigipolisGent\Toerismevlaanderen\Lodging\Request\ListRequest;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @covers \DigipolisGent\Toerismevlaanderen\Lodging\Request\ListRequest
+ */
+class ListRequestTest extends TestCase
+{
+    /**
+     * The constructor creates the proper URI.
+     *
+     * @test
+     */
+    public function uriContainsQuery(): void
+    {
+        $request = new ListRequest(
+            new LocalityFilter('Gent'),
+            new RegistrationStatusFilter('Erkend', 'Vergund')
+        );
+        $expected = 'query=PREFIX+tvl%3A+%3Chttps%3A%2F%2Fdata.vlaanderen.be%2Fns%2Flogies%23%3E+PREFIX+skos%3A+%3Chttp%3A%2F%2Fwww.w3.org%2F2004%2F02%2Fskos%2Fcore%23%3E+PREFIX+tva%3A+%3Chttps%3A%2F%2Fdata.vlaanderen.be%2Fns%2Fadres%23%3E+PREFIX+schema%3A+%3Chttp%3A%2F%2Fschema.org%2F%3E+SELECT+%3F_lodging+%3Fnaam+WHERE+%7B+%3F_lodging+a+tvl%3ALogies%3B+tvl%3AonthaalAdres%2Ftva%3Agemeentenaam+%3Flocality%3B+tvl%3AheeftRegistratie%2Ftvl%3AregistratieStatus%2Fskos%3AprefLabel+%3FregistrationStatus%3B+schema%3Aname+%3Fnaam.+FILTER+%28%3Flocality+%3D+%22Gent%22%40nl%29+FILTER+%28%3FregistrationStatus+%3D+%22Erkend%22%40nl+%7C%7C+%3FregistrationStatus+%3D+%22Vergund%22%40nl%29+%7D+ORDER+BY+%3Fnaam';
+        $this->assertSame($expected, (string) $request->getBody());
+    }
+}