+++ CHANGELOG.md
@@ -8,7 +8,8 @@ Initial package.
 
 ### Added
 
-* Added getting the number of lodges by a given locality.
+* Added getting the number of lodges by given filters.
+* Added getting a list of lodges by given filters.
 
 [1.0.0]: https://github.com/digipolisgent/php_package_dg-toerismevlaanderen-lodging/releases/tag/1.0.0
 [Unreleased]: https://github.com/digipolisgent/php_package_dg-toerismevlaanderen-lodging/compare/master...develop